+++ R/stat_pvalue_manual.R
@@ -1,23 +1,22 @@
-#' @include utilities.R
+#' @include utilities.R geom_bracket.R
 #' @importFrom dplyr pull
-#' @importFrom dplyr tibble
 #' @importFrom glue glue
 NULL
 #'Add Manually P-values to a ggplot
 #'
 #'@description Add manually p-values to a ggplot, such as box blots, dot plots
 #'  and stripcharts.
-#'@inheritParams ggsignif::geom_signif
+#'@inheritParams geom_bracket
 #'@param data a data frame containing statitistical test results. The expected
 #'  default format should contain the following columns: \code{group1 | group2 |
 #'  p | y.position | etc}. \code{group1} and \code{group2} are the groups that
 #'  have been compared. \code{p} is the resulting p-value. \code{y.position} is
 #'  the y coordinates of the p-values in the plot.
-#'@param label the column containing the label (e.g.: label = "p"). Default
-#'  value is "p", for p-value. Can be also an expression that can be formatted
-#'  by the \code{\link[glue]{glue}()} package. For example, when specifying
-#'  label = "t-test, p = \{p\}", the expression \{p\} will be replaced by its
-#'  value.
+#'@param label the column containing the label (e.g.: label = "p" or label =
+#'  "p.adj"), where \code{p} is the p-value. Can be also an expression that can
+#'  be formatted by the \code{\link[glue]{glue}()} package. For example, when
+#'  specifying label = "t-test, p = \{p\}", the expression \{p\} will be
+#'  replaced by its value.
 #'@param y.position column containing the coordinates (in data units) to be used
 #'  for absolute positioning of the label. Default value is "y.position". Can be
 #'  also a numeric vector.
@@ -30,13 +29,22 @@ NULL
 #'  p-value as text (without brackets).
 #'@param size,label.size size of label text.
 #'@param bracket.size Width of the lines of the bracket.
+#'@param color text and line color. Can be variable name in the data for coloring by groups.
+#'@param linetype linetype. Can be variable name in the data for changing linetype by groups.
 #'@param tip.length numeric vector with the fraction of total height that the
 #'  bar goes down to indicate the precise column. Default is 0.03.
 #'@param remove.bracket logical, if \code{TRUE}, brackets are removed from the
 #'  plot. Considered only in the situation, where comparisons are performed
 #'  against reference group or against "all".
+#'@param hide.ns logical value. If TRUE, hide ns symbol when displaying
+#'  significance levels. Filter is done by checking the column
+#'  \code{p.adj.signif}, \code{p.signif}, \code{p.adj} and \code{p}.
+#'@param vjust move the text up or down relative to the bracket. Can be also a
+#'  column name available in the data.
 #'@param position position adjustment, either as a string, or the result of a
 #'  call to a position adjustment function.
+#'@param ... other arguments passed to the function \code{geom_bracket()} or
+#'  \code{geom_text()}
 #'@seealso \code{\link{stat_compare_means}}
 #'@examples
 #'
@@ -91,12 +99,20 @@ NULL
 #'
 #'@export
 stat_pvalue_manual <- function(
-  data, label = "p", y.position = "y.position",
+  data, label = NULL, y.position = "y.position",
   xmin = "group1", xmax = "group2", x = NULL,
-  size = 3.88, label.size = size, bracket.size = 0.3, tip.length = 0.03,
-  remove.bracket = FALSE, position = "identity", ...
+  size = 3.88, label.size = size, bracket.size = 0.3,
+  color = "black", linetype = 1, tip.length = 0.03,
+  remove.bracket = FALSE, step.increase = 0, step.group.by = NULL,
+  hide.ns = FALSE, vjust = 0, position = "identity", ...
 )
 {
+  if(is.null(label)){
+    label <- guess_signif_label_column(data)
+  }
+  if(hide.ns){
+    data <- remove_ns(data)
+  }
   asserttat_group_columns_exists(data)
   comparison <- detect_comparison_type(data)
   all.x.is.missing <- is.null(x) & missing(xmin) & missing(xmax)
@@ -104,31 +120,14 @@ stat_pvalue_manual <- function(
     is.grouped <- length(data$group2) > length(unique(data$group2))
     if(!is.grouped) x <- "group2" # labels will be plotted at x = "group2"
   }
-  # detect automatically if xmin and xmax exists in the data.
+  # Detect automatically if xmin and xmax exists in the data.
   if(all.x.is.missing){
     if(all(c("xmin", "xmax") %in% colnames(data))){
       xmin <- "xmin"
       xmax <- "xmax"
     }
   }
-  # If label is a glue package expression
-  if(.contains_curlybracket(label)){
-    data <- data %>% mutate(label = glue(label))
-    label <- "label"
-  }
-  # P-value displayed as text (without brackets)
-  if(!missing(x)){
-    xmin <- x
-    xmax <- NULL
-  }
-
-  available.variables <- colnames(data)
-
-  if(!(label %in% available.variables))
-    stop("can't find the label variable '", label, "' in the data")
-  if(!(xmin %in% available.variables))
-    stop("can't find the xmin variable '", xmin, "' in the data")
-
+  # should stay before (!is.null(x))
   if(remove.bracket){
     group1.length <- unique(data$group1) %>% length()
     if(group1.length == 1) {
@@ -136,10 +135,30 @@ stat_pvalue_manual <- function(
       xmax <- NULL
     }
   }
+  # P-value displayed as text (without brackets)
+  if(!is.null(x)){
+    xmin <- x
+    xmax <- NULL
+  }
+
+  # If label is a glue package expression
+  if(.contains_curlybracket(label)){
+    data <- data %>% mutate(label = glue(label))
+    label <- "label"
+  }
+
+  available.variables <- colnames(data)
+  if(!(label %in% available.variables))
+    stop("can't find the label variable '", label, "' in the data")
+  if(!(xmin %in% available.variables))
+    stop("can't find the xmin variable '", xmin, "' in the data")
 
   y.position <- .valide_y_position(y.position, data)
-  if(is.numeric(y.position))
+  if(is.numeric(y.position)){
     data$y.position <- y.position
+    y.position <- "y.position"
+  }
+
 
   # If xmax is null, pvalue is drawn as text
   if(!is.null(xmax)) {
@@ -151,31 +170,45 @@ stat_pvalue_manual <- function(
     pvalue.geom <- "text"
   }
 
+  if(!is.null(xmin)){
+    xmin <- data %>% pull(!!xmin)
+  }
+  else{
+    xmin <- NA
+  }
+
   # Build the statistical table for plotting
   xxmax <-xmax  # so that mutate will avoid re-using an existing xmax in the data
+  xxmin <- xmin
   data <- data %>%
     dplyr::mutate(
       label = as.character(data %>% pull(!!label)),
-      y.position = data %>% pull(y.position),
-      xmin = data %>% pull(!!xmin),
+      y.position = data %>% pull(!!y.position),
+      xmin = xxmin,
       xmax = xxmax
     )
+  # vjust
+  if(is.character(vjust)){
+    vjust <- data %>% pull(!!vjust)
+  }
+  else if(missing(vjust)){
+    vjust <- guess_labels_default_vjust(data$label)
+  }
+  data <- data %>% mutate(vjust = !!vjust)
+
 
   if(pvalue.geom == "bracket"){
     if(identical(data$xmin, data$xmax) | remove.bracket){
       # case when ref.group = "all"
       bracket.size = 0
     }
-    mapping <- aes(
-      xmin = xmin, xmax = xmax,
-      annotations = label, y_position = y.position,
-      group = 1:nrow(data)
-    )
-    ggsignif::geom_signif(
-      mapping = mapping, data = data,
-      manual= TRUE, tip_length =  tip.length,
-      textsize = label.size, size = bracket.size,
-      position = position, ...
+    geom_exec(
+      geom_bracket, data = data, xmin = "xmin", xmax = "xmax",
+      label = "label", y.position = "y.position", vjust = "vjust",
+      group = 1:nrow(data),  tip.length =  tip.length,
+      label.size = label.size, size = bracket.size, color = color,
+      linetype = linetype, step.increase = step.increase,
+      step.group.by = step.group.by, position = position, ...
     )
   }
   else{
@@ -183,19 +216,21 @@ stat_pvalue_manual <- function(
       ref.group <- unique(data$group1)
       group2 <- NULL
       data <- add_ctr_rows(data, ref.group = ref.group)
-      mapping <- aes(x = xmin, y = y.position, label = label, group = group2)
+      mapping <- aes(x = xmin, y = y.position, vjust = vjust, label = label, group = group2)
       if(missing(position) & !missing(x)){
         if (is_grouping_variable(x))
           position <- position_dodge(0.8)
       }
     }
     else{
-      mapping <- aes(x = xmin, y = y.position, label = label)
+      mapping <- aes(x = xmin, y = y.position, vjust = vjust, label = label)
     }
-    # X axis variable should be a factor
-    #if(!is.factor(data$xmin))
-    #data$xmin <- factor(data$xmin, levels = unique(data$xmin))
-    geom_text(mapping, data = data, size = label.size, position = position, ...)
+
+    option <- list(data = data, size = label.size, position = position, ...)
+    if(color %in% colnames(data)) mapping$colour <- rlang::ensym(color)
+    else option$color <- color
+    option[["mapping"]] <- mapping
+    do.call(geom_text, option)
   }
 }
 
@@ -256,8 +291,12 @@ detect_comparison_type <- function(data){
   else if (ngroup1 >= 2 & ngroup2 >= 2){
     type = "pairwise"
   }
+  else if(c("group1", "group2") %in% colnames(data)){
+    # filtered data
+    type = "pairwise"
+  }
   else{
-    stop("Make sure that group1 and group2 columns exist in the data")
+    stop("Make sure that group1 and group2 columns exist in the data.")
   }
   type
 }
@@ -271,4 +310,49 @@ is_null_model <- function(data){
 
 is_grouping_variable <- function(x){
   !(x %in% c("group1", "group2"))
+}
+
+
+
+# Check if label column contains stars
+contains_signif_stars <- function(data, label.col){
+  result <- FALSE
+  if(label.col[1] %in% colnames(data)){
+    labels <- data %>% pull(!!label.col)
+    stars <- c("****", "***", "**", "*")
+    result <- any(labels %in% stars)
+  }
+  result
+}
+
+# guess label default vjust
+guess_label_default_vjust <- function(label){
+  if(label %in% c("****", "***", "**", "*"))
+    vjust <- 0.5
+  else vjust <- 0
+  vjust
+}
+guess_labels_default_vjust <- function(labels){
+  labels %>%
+    purrr::map(guess_label_default_vjust) %>%
+    unlist()
+}
+
+# remove non significant
+remove_ns <- function(data){
+  filter <- dplyr::filter
+  columns <- colnames(data)
+  if("p.adj.signif" %in% columns){
+    data <- data %>% filter(.data$p.adj.signif != "ns")
+  }
+  else if("p.adj" %in% columns){
+    data <- data %>% filter(.data$p.adj <= 0.05)
+  }
+  else if("p.signif" %in% columns){
+    data <- data %>% filter(.data$p.signif != "ns")
+  }
+  else if("p" %in% columns){
+    data <- data %>% filter(.data$p <= 0.05)
+  }
+  data
 }